Shows how to do the titantic exercise with Hamilton
I believe that the Hamilton code shown is simpler and easier to manage
than what the end result is in step 24. This is a PR to help start
some healthy discussion in the CQ4DS discord.

Why do I think this is better?

1. It's closer to the original notebook code. So you could probably write
Hamilton code from the beginning... or have a light rewrite to make your
code production grade.
2. It's straightforward to follow the flow of execution (and visualize it).
3. You don't need strong OOP skills to make sure the code
does not devolve into a coupled mess. You just need to think in terms
of logic, encapsulate it in a function, and then declare the right dependencies.

It's not perfect code though:

1. There are different, and likely better ways to construct the dataflow
based on actual/expected changes and uses.
2. I was lazy and didn't document/add many tests. But reproducibility is
quite straightforward.

See the README.md for a few more details.

Otherwise, please roast!